// src/app/models/userSelections.model.ts
export class userSelections {
    id: number = 0;
    user_id: number = 0;
    firstName?: string;
    lastName?: string;
    selection_1_id: number = 0;
    selection_1_name?: string;
    selection_2_id: number = 0;
    selection_2_name?: string;
    selection_3_id: number = 0;
    selection_3_name?: string;
    selection_4_id: number = 0;
    selection_4_name?: string;
    selection_5_id: number = 0;
    selection_5_name?: string;
    selection_6_id: number = 0;
    selection_6_name?: string;
    selection_7_id: number = 0;
    selection_7_name?: string;
    selection_8_id: number = 0;
    selection_8_name?: string;
    selection_9_id: number = 0;
    selection_9_name?: string;
    selection_10_id: number = 0;
    selection_10_name?: string;
    total_point?: number;
    total_point2?: number;
    contestRegistered: number = 1;
} 